# Improve runtime performance of --reindex