@LarryRuane 2019-09-29
Running `bitcoind --reindex` causes the client to reread the `.bitcoin/blocks/blknnnnn.dat` files and re-derive all the index and chainstate information. It's similar to IBD but the client uses these local files rather than download blocks from the network. As `LoadExternalBlockFile()` reads each block from disk, it checks to see if its parent has already been seen (and processed). If so, then this block can be immediately processed (by giving it to `AcceptBlock()`). If this block's parent hasn't yet been seen, it cannot be processed until after its parent is seen. Its prev-hash (parent's hash) is put into a map of "pending" blocks (called `mapBlocksUnknownParent`). Later, when its parent is encountered and processed, this map is consulted to see it has a child that's already been seen that can now be processed. If so, `LoadExternalBlockFile()` now has to go back and re-read the child block from disk (`ReadBlockFromDisk()`), deserialize it again, and then process (pass it to `AcceptBlock()`).

Performance isn't very good because about 90% of blocks are read before their parents are read. (I think this is a result of the headers-first IBD performance enhancement of a few years ago.) This can be seen by running `bitcoind --reindex --debug=reindex`, causing many messages like:
```
2019-09-29T00:29:00Z LoadExternalBlockFile: Out of order block 00000000a2268fb2d4ddc0408fae961a96668db58de415d3c9e0a7694eeb1657, parent 000000005fb53da5c3cda78aef28cbc44fde3e435f1dc2ac04df53387049171f not known
```
then, a short time later:
```
2019-09-29T00:29:00Z LoadExternalBlockFile: Processing out of order child 00000000a2268fb2d4ddc0408fae961a96668db58de415d3c9e0a7694eeb1657 of 000000005fb53da5c3cda78aef28cbc44fde3e435f1dc2ac04df53387049171f
```
So 90% of blocks end up being read from disk twice (actually, not twice, since the operating system's buffer cache likely still has the data, but there is still a memory-to-memory copy), and deserialized twice.

This PR mitigates this problem by taking advantage of the fact that recent blocks are still in the buffer (`blkdat`) that is used to read from the `blk` files. Since rewinding to an earlier offset in the buffer has been fixed by the recently merged #16577, we can see if the child block is still in memory (the probability is increased by making the buffer somewhat larger), and if so, avoid having to read the child block data from disk. It turns out that the child block is still in the buffer 60% of the time. This depends on some randomness that's introduced during IBD, so it may be different for you. (My IBD was done using all the default settings, nothing special.)

The second part of this change initially deserializes only the block header, rather than the entire block. since about 90% of the time we're going to have to process this block later; only the `hashPrevBlock` field of the header is needed to determine if we can process this block immediately (if this block's parent has been seen). This does mean that 10% of the time, we deserialize the header, and then immediately also the full block (where it would have been more efficient to simply deserialize the full block), but overall this is a big performance win.

On my system, this PR reduces the time to `--reindex` by 48 minutes. I encourage reviewers to attempt to replicate my findings. Reindexing has two parts, reading the `blocks/blknnnnn.dat` files (which this PR improves), and then connecting the blocks to the blockchain (the `UpdateTip` messages), which this PR doesn't affect. For me, this PR reduces the first part from 123 to 75 minutes. After you start seeing `UpdateTip` messages, you can stop the client, and you will have the time measurement you need from `debug.log`.