@LarryRuane 2019-09-30
I forgot to update the `CBufferedFile` unit tests, just did that and another force-push.