@LarryRuane 2019-10-11
Testnet performance isn't important, but as a sanity check, I verified that the new code runs faster there too on my laptop. The time between `Reindexing block file blk00001.dat...` and `Reindexing finished` reduced from 10m21s to 7m13s.