@LarryRuane 2019-10-15
I force-pushed a reorganization of the commits to make reviewing easier. I split the largest commit, **speed up reindex by reading out-of-order blocks from memory**, into two separate commits:
1. Commit message: **initially deserialize only CBlockHeader, not CBlock**. Deserialize only a block header (only 32 bytes) initially, instead of a full block. The block header is all that's needed to tell us if this block's parent has been seen yet; if not, the block will have to be deserialized later, after its parent is found. Nearly all blocks (89% on my system) are seen before their parents, so this is a big improvement.
1. Commit message: **speed up reindex by reading out-of-order blocks from memory**. If we've just processed a block that's the parent of an earlier-seen block, we must now process that earlier block. Instead of always reading that block from disk, see if it happens to still be in the circular disk read memory buffer ('CBufferedFile blkdat`). This saves a disk read.

We could merge the first commit without the second. The first commit is pretty simple; the second commit is a little more involved.

These two commits have a buffer size unchanged from master, 8mb (`2*MAX_BLOCK_SERIALIZED_SIZE`), so there is no change in memory usage. But with this buffer size, only 29% of the out-of-order blocks are found in the buffer. If we increase the buffer by not too much (as I did in an earlier commit that's now been erased by force-push, 640mb), to 32mb, the "hit rate" increases to 50%. This is probably a good space-time balance.

Here are the block-load times on my system (with all default settings), which is just the first part of reindexing (each row builds on the previous row):

| version | real time (minutes) |
| ----------- | ----------- |
| master | 130 |
| deserialize only header | 90 |
| look for blocks in buffer, 8mb |84 |
| look for blocks in buffer, 32mb |78 |
| look for blocks in buffer, 640mb |72 |

(Take with a grain of salt, because the times varied even on identical runs, and a lot probably depends on the specifics of my system.)