@MarcoFalke 2019-10-18
Under what circumstances is a `-reindex` needed these days? Only when disabling `-prune`? Though, then you'll download from the network anyway.