@kallewoof 2020-03-07
* [x] Benchmarks running on a50f956e304adc7c75428ec391f1a76a99c1ddde:
```
25.11s user 163.37s system 7% cpu 39:53.51 total
25.59s user 162.06s system 7% cpu 39:43.73 total
25.35s user 161.31s system 7% cpu 39:16.82 total
24.98s user 162.59s system 7% cpu 39:06.45 total
25.36s user 162.80s system 8% cpu 39:01.81 total
```

* [ ] With SCHED_BATCH off:
* [ ] Master: